Made a (crude) block highlighter
Works as intended, but is missing some features, like "seeing through"
the block when looking at a section of it that doesn't contain any faces
being rendered, and highlighting when looking at a point on the inside
and close to the edge of the block.

Also did some minor clean-ups and optimisations.