Smoother block highlighter + debug info
When looking at the edge of a redstone paste segment and more than one
face is hit, the face that is closest to the point hit according to an
imagined diagonal line separating the two faces, will now be
highlighted.

Also made some debug info magically appear on the debug screen when
targeting a redstone paste block.