Expanded pasting system + huge code cleanup
Expanded and simplified the pasting system so that each redstone paste
block now supports pasting of up to five sides (six sides is
fundamentally impossible for the player to paste anyway) and stores it
in its metadata.

Also temporarily disabled proper block rendering and fixed a bug that
hindered the placement of neighbour blocks while targeting a redstone
paste block.