Adaptive redstone particles + error fixes
Powered redstone paste now spawns redstone particles near each pasted
side to subtly show which sides a redstone paste block is pasted on and
which sides aren't. Well, it's mostly for aesthetics, but I made it so
it provides *some* level of functionality.

Also, fixed a couple of bugs that made some block update procedures
fail, most noticeably causing both the affected block and the
neighbouring blocks to not render correctly.