Adaptive indirect pasting
When indirectly placing redstone paste, i.e. clicking on a block that
can't be pasted on making the redstone paste being pasted on a
neighbouring block, the side pasted will now be decided by a combination
of the player's position relative to the clicked block and the player's
camera axes. The same happens when pasting an additional side to the
opposite of the side already pasted, requiring a third side to connect
the two.

Also, did a little bit of clean-up and prettification, yw.