Improved power updating on chunk load
I noticed a bug when loading a map after having placed a bunch of
redstone paste blocks in it with MCEdit; their power didn't update on
chunk load. I figured that it was probably caused by MCEdit not adding
the tile entities of the placed blocks to each chunk's tile entity map,
so I made my chunk load handler iterate through every block in every
non-empty sub-chunk instead of iterating through every tile entity in
every chunk. It's a bit heavier, but the vast majority of the time it
only happens on world load, so it won't be an impact on general
gameplay. I mean, when loading a world with the render distance set to
15 chunks, the game only went through line 42 and 44 in
EventHookContainer some 35 000 000 times, so it doesn't look too bad.
Plus it works, now!