Non-targetable generated sides
If a player is looking at a generated side, i.e. a side that is part of
the model, but is not pasted, while sneaking, it will not be targeted.
This allows for easier checks of sides being pasted or not, and players
will now be spared of ending up in situations where they find themselves
unable to remove a side, like it used to be.